+++ cmd/build.js
@@ -15,7 +15,7 @@ shell = require('shelljs');
 var env = JSON.parse(process.env.stagejs);
 
 if(!env['stagejs-version']){
-	console.error('You don\'t have a project here...'.red, 'use \'init\' first.'.yellow);
+	console.error('You don\'t have a project here...'.red, 'run \'stagejs init\' first.'.yellow);
 	process.exit(1);
 }